--- sites/docs/src/routes/(app)/examples/tasks/(components)/data-table-row-actions.svelte
@@ -1,50 +1,0 @@
-<script lang="ts" module>
-	type TData = unknown;
-</script>
-
-<script lang="ts" generics="TData">
-	import EllipsisIcon from "@lucide/svelte/icons/ellipsis";
-	import type { Row } from "@tanstack/table-core";
-	import { labels } from "../(data)/data.js";
-	import { taskSchema } from "../(data)/schemas.js";
-	import * as DropdownMenu from "$lib/registry/ui/dropdown-menu/index.js";
-	import Button from "$lib/registry/ui/button/button.svelte";
-
-	let { row }: { row: Row<TData> } = $props();
-
-	const task = taskSchema.parse(row.original);
-</script>
-
-<DropdownMenu.Root>
-	<DropdownMenu.Trigger>
-		{#snippet child({ props })}
-			<Button {...props} variant="ghost" class="data-[state=open]:bg-muted flex h-8 w-8 p-0">
-				<EllipsisIcon />
-				<span class="sr-only">Open Menu</span>
-			</Button>
-		{/snippet}
-	</DropdownMenu.Trigger>
-	<DropdownMenu.Content class="w-[160px]" align="end">
-		<DropdownMenu.Item>Edit</DropdownMenu.Item>
-		<DropdownMenu.Item>Make a copy</DropdownMenu.Item>
-		<DropdownMenu.Item>Favorite</DropdownMenu.Item>
-		<DropdownMenu.Separator />
-		<DropdownMenu.Sub>
-			<DropdownMenu.SubTrigger>Labels</DropdownMenu.SubTrigger>
-			<DropdownMenu.SubContent>
-				<DropdownMenu.RadioGroup value={task.label}>
-					{#each labels as label (label.value)}
-						<DropdownMenu.RadioItem value={label.value}>
-							{label.label}
-						</DropdownMenu.RadioItem>
-					{/each}
-				</DropdownMenu.RadioGroup>
-			</DropdownMenu.SubContent>
-		</DropdownMenu.Sub>
-		<DropdownMenu.Separator />
-		<DropdownMenu.Item>
-			Delete
-			<DropdownMenu.Shortcut>⌘⌫</DropdownMenu.Shortcut>
-		</DropdownMenu.Item>
-	</DropdownMenu.Content>
-</DropdownMenu.Root>
