--- sites/docs/src/routes/(app)/examples/tasks/(components)/index.ts
@@ -1,10 +1,0 @@
-export { default as DataTableCheckbox } from "./data-table-checkbox.svelte";
-export { default as DataTableTitleCell } from "./data-table-title-cell.svelte";
-export { default as DataTableStatusCell } from "./data-table-status-cell.svelte";
-export { default as DataTableRowActions } from "./data-table-row-actions.svelte";
-export { default as DataTablePriorityCell } from "./data-table-priority-cell.svelte";
-export { default as DataTableColumnHeader } from "./data-table-column-header.svelte";
-export { default as DataTableToolbar } from "./data-table-toolbar.svelte";
-export { default as DataTablePagination } from "./data-table-pagination.svelte";
-export { default as DataTableViewOptions } from "./data-table-view-options.svelte";
-export { default as DataTableFacetedFilter } from "./data-table-faceted-filter.svelte";
