+++ sites/docs/src/routes/(app)/examples/tasks/components/data-table-faceted-filter.svelte
