+++ sites/docs/src/routes/(app)/examples/tasks/components/data-table-toolbar.svelte
@@ -5,10 +5,10 @@
 <script lang="ts" generics="TData">
 	import XIcon from "@lucide/svelte/icons/x";
 	import type { Table } from "@tanstack/table-core";
-	import { priorities, statuses } from "../(data)/data.js";
 	import { DataTableFacetedFilter, DataTableViewOptions } from "./index.js";
 	import Button from "$lib/registry/ui/button/button.svelte";
 	import { Input } from "$lib/registry/ui/input/index.js";
+	import { priorities, statuses } from "../data/data.js";
 
 	let { table }: { table: Table<TData> } = $props();
