+++ sites/docs/src/routes/(app)/examples/tasks/components/data-table-view-options.svelte
