+++ sites/docs/src/routes/(app)/examples/tasks/components/data-table.svelte
@@ -1,0 +1,456 @@
+<script lang="ts">
+	import {
+		type ColumnDef,
+		type ColumnFiltersState,
+		type PaginationState,
+		type Row,
+		type RowSelectionState,
+		type SortingState,
+		type VisibilityState,
+		type Table as TableType,
+		getCoreRowModel,
+		getFacetedRowModel,
+		getFacetedUniqueValues,
+		getFilteredRowModel,
+		getPaginationRowModel,
+		getSortedRowModel,
+		type Column,
+	} from "@tanstack/table-core";
+	import DataTableToolbar from "./data-table-toolbar.svelte";
+	import { createSvelteTable } from "$lib/registry/ui/data-table/data-table.svelte.js";
+	import FlexRender from "$lib/registry/ui/data-table/flex-render.svelte";
+	import * as Table from "$lib/registry/ui/table/index.js";
+	import * as DropdownMenu from "$lib/registry/ui/dropdown-menu/index.js";
+	import { labels, priorities, statuses } from "../data/data.js";
+	import { taskSchema, type Task } from "../data/schemas.js";
+	import { renderComponent, renderSnippet } from "$lib/registry/ui/data-table/render-helpers.js";
+	import Checkbox from "$lib/registry/ui/checkbox/checkbox.svelte";
+	import { createRawSnippet } from "svelte";
+	import { Badge } from "$lib/registry/ui/badge/index.js";
+	import { Button } from "$lib/registry/ui/button/index.js";
+	import EllipsisIcon from "@lucide/svelte/icons/ellipsis";
+	import ChevronRightIcon from "@lucide/svelte/icons/chevron-right";
+	import ChevronLeftIcon from "@lucide/svelte/icons/chevron-left";
+	import ChevronsLeftIcon from "@lucide/svelte/icons/chevrons-left";
+	import ChevronsRightIcon from "@lucide/svelte/icons/chevrons-right";
+	import ArrowUpIcon from "@lucide/svelte/icons/arrow-up";
+	import ArrowDownIcon from "@lucide/svelte/icons/arrow-down";
+	import ChevronsUpDownIcon from "@lucide/svelte/icons/chevrons-up-down";
+	import EyeOffIcon from "@lucide/svelte/icons/eye-off";
+	import * as Select from "$lib/registry/ui/select/index.js";
+	import type { HTMLAttributes } from "svelte/elements";
+	import { cn } from "$lib/utils.js";
+
+	let { data }: { data: Task[] } = $props();
+
+	let rowSelection = $state<RowSelectionState>({});
+	let columnVisibility = $state<VisibilityState>({});
+	let columnFilters = $state<ColumnFiltersState>([]);
+	let sorting = $state<SortingState>([]);
+	let pagination = $state<PaginationState>({ pageIndex: 0, pageSize: 10 });
+
+	const columns: ColumnDef<Task>[] = [
+		{
+			id: "select",
+			header: ({ table }) =>
+				renderComponent(Checkbox, {
+					checked: table.getIsAllPageRowsSelected(),
+					onCheckedChange: (value) => table.toggleAllPageRowsSelected(value),
+					indeterminate:
+						table.getIsSomePageRowsSelected() && !table.getIsAllPageRowsSelected(),
+					"aria-label": "Select all",
+					class: "translate-y-[2px]",
+				}),
+			cell: ({ row }) =>
+				renderComponent(Checkbox, {
+					checked: row.getIsSelected(),
+					onCheckedChange: (value) => row.toggleSelected(value),
+					"aria-label": "Select row",
+					class: "translate-y-[2px]",
+				}),
+			enableSorting: false,
+			enableHiding: false,
+		},
+		{
+			accessorKey: "id",
+			header: ({ column }) => {
+				return renderSnippet(ColumnHeader, {
+					column,
+					title: "Task",
+				});
+			},
+			cell: ({ row }) => {
+				const idSnippet = createRawSnippet<[string]>((getId) => {
+					const id = getId();
+					return {
+						render: () => `<div class="w-[80px]">${id}</div>`,
+					};
+				});
+
+				return renderSnippet(idSnippet, row.getValue("id"));
+			},
+			enableSorting: false,
+			enableHiding: false,
+		},
+		{
+			accessorKey: "title",
+			header: ({ column }) => renderSnippet(ColumnHeader, { column, title: "Title" }),
+			cell: ({ row }) => {
+				return renderSnippet(TitleCell, {
+					labelValue: row.original.label,
+					value: row.original.title,
+				});
+			},
+		},
+		{
+			accessorKey: "status",
+			header: ({ column }) =>
+				renderSnippet(ColumnHeader, {
+					column,
+					title: "Status",
+				}),
+			cell: ({ row }) => {
+				return renderSnippet(StatusCell, {
+					value: row.original.status,
+				});
+			},
+			filterFn: (row, id, value) => {
+				return value.includes(row.getValue(id));
+			},
+		},
+		{
+			accessorKey: "priority",
+			header: ({ column }) => {
+				return renderSnippet(ColumnHeader, {
+					title: "Priority",
+					column,
+				});
+			},
+			cell: ({ row }) => {
+				return renderSnippet(PriorityCell, {
+					value: row.original.priority,
+				});
+			},
+			filterFn: (row, id, value) => {
+				return value.includes(row.getValue(id));
+			},
+		},
+		{
+			id: "actions",
+			cell: ({ row }) => renderSnippet(RowActions, { row }),
+		},
+	];
+
+	const table = createSvelteTable({
+		get data() {
+			return data;
+		},
+		state: {
+			get sorting() {
+				return sorting;
+			},
+			get columnVisibility() {
+				return columnVisibility;
+			},
+			get rowSelection() {
+				return rowSelection;
+			},
+			get columnFilters() {
+				return columnFilters;
+			},
+			get pagination() {
+				return pagination;
+			},
+		},
+		columns,
+		enableRowSelection: true,
+		onRowSelectionChange: (updater) => {
+			if (typeof updater === "function") {
+				rowSelection = updater(rowSelection);
+			} else {
+				rowSelection = updater;
+			}
+		},
+		onSortingChange: (updater) => {
+			if (typeof updater === "function") {
+				sorting = updater(sorting);
+			} else {
+				sorting = updater;
+			}
+		},
+		onColumnFiltersChange: (updater) => {
+			if (typeof updater === "function") {
+				columnFilters = updater(columnFilters);
+			} else {
+				columnFilters = updater;
+			}
+		},
+		onColumnVisibilityChange: (updater) => {
+			if (typeof updater === "function") {
+				columnVisibility = updater(columnVisibility);
+			} else {
+				columnVisibility = updater;
+			}
+		},
+		onPaginationChange: (updater) => {
+			if (typeof updater === "function") {
+				pagination = updater(pagination);
+			} else {
+				pagination = updater;
+			}
+		},
+		getCoreRowModel: getCoreRowModel(),
+		getFilteredRowModel: getFilteredRowModel(),
+		getPaginationRowModel: getPaginationRowModel(),
+		getSortedRowModel: getSortedRowModel(),
+		getFacetedRowModel: getFacetedRowModel(),
+		getFacetedUniqueValues: getFacetedUniqueValues(),
+	});
+</script>
+
+{#snippet StatusCell({ value }: { value: string })}
+	{@const status = statuses.find((status) => status.value === value)}
+	{#if status}
+		<div class="flex w-[100px] items-center">
+			<status.icon class="text-muted-foreground mr-2 size-4" />
+			<span>{status.label}</span>
+		</div>
+	{/if}
+{/snippet}
+
+{#snippet TitleCell({ value, labelValue }: { value: string; labelValue: string })}
+	{@const label = labels.find((label) => label.value === labelValue)}
+	<div class="flex space-x-2">
+		{#if label}
+			<Badge variant="outline">{label.label}</Badge>
+		{/if}
+		<span class="max-w-[500px] truncate font-medium">
+			{value}
+		</span>
+	</div>
+{/snippet}
+
+{#snippet PriorityCell({ value }: { value: string })}
+	{@const priority = priorities.find((priority) => priority.value === value)}
+	{#if priority}
+		<div class="flex items-center">
+			<priority.icon class="text-muted-foreground mr-2 size-4" />
+			<span>{priority.label}</span>
+		</div>
+	{/if}
+{/snippet}
+
+{#snippet RowActions({ row }: { row: Row<Task> })}
+	{@const task = taskSchema.parse(row.original)}
+	<DropdownMenu.Root>
+		<DropdownMenu.Trigger>
+			{#snippet child({ props })}
+				<Button
+					{...props}
+					variant="ghost"
+					class="data-[state=open]:bg-muted flex h-8 w-8 p-0"
+				>
+					<EllipsisIcon />
+					<span class="sr-only">Open Menu</span>
+				</Button>
+			{/snippet}
+		</DropdownMenu.Trigger>
+		<DropdownMenu.Content class="w-[160px]" align="end">
+			<DropdownMenu.Item>Edit</DropdownMenu.Item>
+			<DropdownMenu.Item>Make a copy</DropdownMenu.Item>
+			<DropdownMenu.Item>Favorite</DropdownMenu.Item>
+			<DropdownMenu.Separator />
+			<DropdownMenu.Sub>
+				<DropdownMenu.SubTrigger>Labels</DropdownMenu.SubTrigger>
+				<DropdownMenu.SubContent>
+					<DropdownMenu.RadioGroup value={task.label}>
+						{#each labels as label (label.value)}
+							<DropdownMenu.RadioItem value={label.value}>
+								{label.label}
+							</DropdownMenu.RadioItem>
+						{/each}
+					</DropdownMenu.RadioGroup>
+				</DropdownMenu.SubContent>
+			</DropdownMenu.Sub>
+			<DropdownMenu.Separator />
+			<DropdownMenu.Item>
+				Delete
+				<DropdownMenu.Shortcut>⌘⌫</DropdownMenu.Shortcut>
+			</DropdownMenu.Item>
+		</DropdownMenu.Content>
+	</DropdownMenu.Root>
+{/snippet}
+
+{#snippet Pagination({ table }: { table: TableType<Task> })}
+	<div class="flex items-center justify-between px-2">
+		<div class="text-muted-foreground flex-1 text-sm">
+			{table.getFilteredSelectedRowModel().rows.length} of
+			{table.getFilteredRowModel().rows.length} row(s) selected.
+		</div>
+		<div class="flex items-center space-x-6 lg:space-x-8">
+			<div class="flex items-center space-x-2">
+				<p class="text-sm font-medium">Rows per page</p>
+				<Select.Root
+					allowDeselect={false}
+					type="single"
+					value={`${table.getState().pagination.pageSize}`}
+					onValueChange={(value) => {
+						table.setPageSize(Number(value));
+					}}
+				>
+					<Select.Trigger class="h-8 w-[70px]">
+						{String(table.getState().pagination.pageSize)}
+					</Select.Trigger>
+					<Select.Content side="top">
+						{#each [10, 20, 30, 40, 50] as pageSize (pageSize)}
+							<Select.Item value={`${pageSize}`}>
+								{pageSize}
+							</Select.Item>
+						{/each}
+					</Select.Content>
+				</Select.Root>
+			</div>
+			<div class="flex w-[100px] items-center justify-center text-sm font-medium">
+				Page {table.getState().pagination.pageIndex + 1} of
+				{table.getPageCount()}
+			</div>
+			<div class="flex items-center space-x-2">
+				<Button
+					variant="outline"
+					class="hidden size-8 p-0 lg:flex"
+					onclick={() => table.setPageIndex(0)}
+					disabled={!table.getCanPreviousPage()}
+				>
+					<span class="sr-only">Go to first page</span>
+					<ChevronsLeftIcon />
+				</Button>
+				<Button
+					variant="outline"
+					class="size-8 p-0"
+					onclick={() => table.previousPage()}
+					disabled={!table.getCanPreviousPage()}
+				>
+					<span class="sr-only">Go to previous page</span>
+					<ChevronLeftIcon />
+				</Button>
+				<Button
+					variant="outline"
+					class="size-8 p-0"
+					onclick={() => table.nextPage()}
+					disabled={!table.getCanNextPage()}
+				>
+					<span class="sr-only">Go to next page</span>
+					<ChevronRightIcon />
+				</Button>
+				<Button
+					variant="outline"
+					class="hidden size-8 p-0 lg:flex"
+					onclick={() => table.setPageIndex(table.getPageCount() - 1)}
+					disabled={!table.getCanNextPage()}
+				>
+					<span class="sr-only">Go to last page</span>
+					<ChevronsRightIcon />
+				</Button>
+			</div>
+		</div>
+	</div>
+{/snippet}
+
+{#snippet ColumnHeader({
+	column,
+	title,
+	class: className,
+	...restProps
+}: { column: Column<Task>; title: string } & HTMLAttributes<HTMLDivElement>)}
+	{#if !column?.getCanSort()}
+		<div class={className} {...restProps}>
+			{title}
+		</div>
+	{:else}
+		<div class={cn("flex items-center", className)} {...restProps}>
+			<DropdownMenu.Root>
+				<DropdownMenu.Trigger>
+					{#snippet child({ props })}
+						<Button
+							{...props}
+							variant="ghost"
+							size="sm"
+							class="data-[state=open]:bg-accent -ml-3 h-8"
+						>
+							<span>
+								{title}
+							</span>
+							{#if column.getIsSorted() === "desc"}
+								<ArrowDownIcon />
+							{:else if column.getIsSorted() === "asc"}
+								<ArrowUpIcon />
+							{:else}
+								<ChevronsUpDownIcon />
+							{/if}
+						</Button>
+					{/snippet}
+				</DropdownMenu.Trigger>
+				<DropdownMenu.Content align="start">
+					<DropdownMenu.Item onclick={() => column.toggleSorting(false)}>
+						<ArrowUpIcon class="text-muted-foreground/70 mr-2 size-3.5" />
+						Asc
+					</DropdownMenu.Item>
+					<DropdownMenu.Item onclick={() => column.toggleSorting(true)}>
+						<ArrowDownIcon class="text-muted-foreground/70 mr-2 size-3.5" />
+						Desc
+					</DropdownMenu.Item>
+					<DropdownMenu.Separator />
+					<DropdownMenu.Item onclick={() => column.toggleVisibility(false)}>
+						<EyeOffIcon class="text-muted-foreground/70 mr-2 size-3.5" />
+						Hide
+					</DropdownMenu.Item>
+				</DropdownMenu.Content>
+			</DropdownMenu.Root>
+		</div>
+	{/if}
+{/snippet}
+
+<div class="space-y-4">
+	<DataTableToolbar {table} />
+	<div class="rounded-md border">
+		<Table.Root>
+			<Table.Header>
+				{#each table.getHeaderGroups() as headerGroup (headerGroup.id)}
+					<Table.Row>
+						{#each headerGroup.headers as header (header.id)}
+							<Table.Head colspan={header.colSpan}>
+								{#if !header.isPlaceholder}
+									<FlexRender
+										content={header.column.columnDef.header}
+										context={header.getContext()}
+									/>
+								{/if}
+							</Table.Head>
+						{/each}
+					</Table.Row>
+				{/each}
+			</Table.Header>
+			<Table.Body>
+				{#each table.getRowModel().rows as row (row.id)}
+					<Table.Row data-state={row.getIsSelected() && "selected"}>
+						{#each row.getVisibleCells() as cell (cell.id)}
+							<Table.Cell>
+								<FlexRender
+									content={cell.column.columnDef.cell}
+									context={cell.getContext()}
+								/>
+							</Table.Cell>
+						{/each}
+					</Table.Row>
+				{:else}
+					<Table.Row>
+						<Table.Cell colspan={columns.length} class="h-24 text-center">
+							No results.
+						</Table.Cell>
+					</Table.Row>
+				{/each}
+			</Table.Body>
+		</Table.Root>
+	</div>
+	{@render Pagination({ table })}
+</div>
