+++ sites/docs/src/routes/(app)/examples/tasks/components/index.ts
@@ -1,0 +1,3 @@
+export { default as DataTableToolbar } from "./data-table-toolbar.svelte";
+export { default as DataTableViewOptions } from "./data-table-view-options.svelte";
+export { default as DataTableFacetedFilter } from "./data-table-faceted-filter.svelte";
